minor warning
git-svn-id: https://ricamsvn.ricam.oeaw.ac.at/gismo/stable@6486 e128ff1e-fa79-4151-a8cb-522299a2c450

+++ src/gsNurbs/gsBoehm.hpp
@@ -171,7 +171,7 @@ void gsBoehmRefine( KnotVectorType & knots,
 
     for(int j = 0; j <= a; j++)
         nknots[j] = knots[j];
-    for(int j = b+p; j < knots.size(); j++)
+    for(size_t j = b+p; j < knots.size(); j++)
         nknots[j + nk] = knots[j];
 
     int i = b + p - 1;
